chore: partial Implement stage progress (incomplete)
